Tolerate older usage with mandatory routename and optional path
With the following review:
https://github.com/bbangert/routes/commit/0a417004be7e2d950bdcd629ccf24cf9f56ef817

Routes 2.3 has changed the function signature to mandate
path in addition to routename for the connect method.

In this patch, we try to get back to path being optional by
checking the length of the args list. No other change in
logic.

This hopefully fixes the problem reported here:
https://github.com/bbangert/routes/issues/64

Add an entry in Changelog as well.