Specify supersieve version (py2 compatibility)
BeautifulSoup states supersieve > 1.2 which is what gets picked up by
distutils. v2.0 is found and installed. Setup then moves on to the next
package in test_requires and when checking versions discovers the conditional
in BeautifulSoup setup.py that if python major version == 2 then
supersieve < v2.0 should be installed.

If we need to preserve compatibility with py2 then reluctantly specify a
version in our setup.